refactor: move user to package
