--- src/main/java/com/alkemy/ong/domain/categories/User.java
@@ -1,15 +1,0 @@
-package com.alkemy.ong.domain.model;
-
-public class User {
-  String firstName;
-
-  String lastName;
-
-  String email;
-
-  String password;
-
-  String photo;
-
-  Role role;
-}
