+++ src/main/java/com/alkemy/ong/domain/user/User.java
@@ -1,0 +1,17 @@
+package com.alkemy.ong.domain.user;
+
+import com.alkemy.ong.domain.model.Role;
+
+public class User {
+  String firstName;
+
+  String lastName;
+
+  String email;
+
+  String password;
+
+  String photo;
+
+  Role role;
+}
